refactor(healthcheck): move to separate file
