+++ src/index.js
@@ -7,6 +7,7 @@ const AccountAPI = require('./api-account.js');
 const AuthAPI = require('./api-auth.js');
 const PurgeAPI = require('./api-purge.js');
 const DomainAPI = require('./api-domain.js');
+const HealthcheckAPI = require('./api-healthcheck');
 
 class RateLimitError extends Error {
 
@@ -304,7 +305,7 @@ class Fastly {
     this.headers = new Headers(this);
 
     // bind the methods of the API classes.
-    [AccountAPI, AuthAPI, PurgeAPI, DomainAPI].forEach((API) => {
+    [AccountAPI, AuthAPI, PurgeAPI, DomainAPI, HealthcheckAPI].forEach((API) => {
       const api = new API(this);
       Object.getOwnPropertyNames(API.prototype).forEach((name) => {
         const prop = api[name];
@@ -1235,61 +1236,6 @@ class Fastly {
    */
   async dryrun(operations) {
     return this.transact(operations, false);
-  }
-
-  /**
-   * List all healthchecks for a particular service and version.
-   *
-   * @param {string} version - The current version of a service.
-   * @returns {Promise} The response object representing the completion or failure.
-   */
-  async readHealthchecks(version) {
-    return this.request.get(`/service/${this.service_id}/version/${await this.getVersion(version, 'current')}/healthcheck`);
-  }
-
-  /**
-   * Get details of a single named healthcheck.
-   *
-   * @param {string} version - The current version of a service.
-   * @param {string} name - The name of the healthcheck.
-   * @returns {Promise} The response object representing the completion or failure.
-   */
-  async readHealthcheck(version, name) {
-    return this.request.get(`/service/${this.service_id}/version/${await this.getVersion(version, 'current')}/healthcheck/${name}`);
-  }
-
-  /**
-   * Create a healthcheck for a particular service and version.
-   *
-   * @param {string} version - The current version of a service.
-   * @param {object} data - The healthcheck definition.
-   * @returns {Promise} The response object representing the completion or failure.
-   */
-  async createHealthcheck(version, data) {
-    return this.request.post(`/service/${this.service_id}/version/${await this.getVersion(version, 'current')}/healthcheck`, data);
-  }
-
-  /**
-   * Update the healthcheck for a particular service and version.
-   *
-   * @param {string} version - The current version of a service.
-   * @param {string} name - The name of the healthcheck to update.
-   * @param {object} data - The healthcheck definition.
-   * @returns {Promise} The response object representing the completion or failure.
-   */
-  async updateHealthcheck(version, name, data) {
-    return this.request.put(`/service/${this.service_id}/version/${await this.getVersion(version, 'current')}/healthcheck/${name}`, data);
-  }
-
-  /**
-   * Delete the healthcheck for a particular service and version.
-   *
-   * @param {string} version - The current version of a service.
-   * @param {string} name - The name of the healthcheck to delete.
-   * @returns {Promise} The response object representing the completion or failure.
-   */
-  async deleteHealthcheck(version, name) {
-    return this.request.delete(`/service/${this.service_id}/version/${await this.getVersion(version, 'current')}/healthcheck/${name}`);
   }
 }
